CORE-4750 Fix for Analysis search JSON as URI query.
The URI#getQuery method was not decoding query parameters correctly, so that the filter parameter JSON array was initially decoded with '+' replacing spaces inside it, which caused the rest of the array contents to not be decoded by the time the proxyToken parameter was appended.
Updated BaseUrlConnector#addQueryParam to replace the URI#getQuery method call with a getRawQuery call, and to use URLDecoder#decode to properly decode the original query before appending the new param.